demo/word_analysis_system/requirements.txt
google-generativeai
python-dotenv
tqdm
postgrest-py>=0.7.0
supabase-py>=1.0.0
